requirements.txt
jsonpath-rw-ext==1.0.0
jsonschema==2.5.1
requests==2.12.4
robotframework-databaselibrary==0.8.1
robotframework-httplibrary==0.4.2
robotframework-lint==0.9
robotframework-pabot==0.44
robotframework-rammbock==0.4.0.1 ; python_version <= '2.7'
robotframework-requests==0.4.6
robotframework-seleniumlibrary==3.0.1
robotframework-sshlibrary==2.1.3 ; python_version <= '2.7'
robotframework==3.0.1
robotremoteserver==1.1
stopit
